craft/requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiosignal==1.3.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
cx_Freeze==7.0.0
cx_Logging==3.2.0
filelock==3.9.0
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2023.4.0
grpcio==1.62.1
idna==3.6
intel-openmp==2021.4.0
Jinja2==3.1.2
lief==0.14.1
lightning-utilities==0.11.2
Markdown==3.6
MarkupSafe==2.1.3
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.2
pillow==10.3.0
protobuf==5.26.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
six==1.16.0
sympy==1.12
tbb==2021.11.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
torch==2.3.0
torch-tb-profiler==0.4.3
torchmetrics==1.3.2
torchvision==0.18.0
tqdm==4.66.4
ttkbootstrap==1.10.1
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
yarl==1.9.4
